Move azure client logic from AzureStorageService to AzureBlobStore
